# Conflicts:
#	scripts/chce_nextcloud.sh

+++ scripts/chce_nextcloud.sh
@@ -1,18 +1,18 @@
 #!/bin/bash
 # NEXTCLOUD installation script
-# Autor: Mariusz 'maniek205' Kowalski
+# Autors: Mariusz 'maniek205' Kowalski, Marcin Wozniak
 
 USERNAME=admin
-PASSWORD=$(head -c 100 /dev/urandom | tr -dc A-Za-z0-9 | head -c13) 
+PASSWORD=$(head -c 100 /dev/urandom | tr -dc A-Za-z0-9 | head -c13)
 
 DB_USER=root
-DB_PASS=$(head -c 100 /dev/urandom | tr -dc A-Za-z0-9 | head -c13) 
+DB_PASS=$(head -c 100 /dev/urandom | tr -dc A-Za-z0-9 | head -c13)
 
 NEXT_CLOUD_USER=admin
 NEXT_CLOUD_PASS=$(head -c 100 /dev/urandom | tr -dc A-Za-z0-9 | head -c13)
 
 #Set Timezone to prevent installation interruption
-ln -snf /usr/share/zoneinfo/Poland /etc/localtime && echo "Etc/UTC" > /etc/timezone
+[[ ! -f /etc/localtime ]] && ln -snf /usr/share/zoneinfo/Poland /etc/localtime && echo "Etc/UTC" > /etc/timezone
 
 #Installing prerequisites https://docs.nextcloud.com/server/latest/admin_manual/installation/example_ubuntu.html
 apt update
@@ -22,24 +22,27 @@ apt install -y php7.4-gmp php7.4-bcmath php-imagick php7.4-xml php7.4-zip
 
 #Configuring mariaDB
 /etc/init.d/mysql start
-mysql -u$DB_USER -p$DB_PASS -e "CREATE USER '$USERNAME'@'localhost' IDENTIFIED BY '$PASSWORD'; 
-CREATE DATABASE IF NOT EXISTS nextcloud CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci; 
-GRANT ALL PRIVILEGES ON nextcloud.* TO '$USERNAME'@'localhost'; 
+mysql -u"$DB_USER" -p"$DB_PASS" -e "CREATE USER '$USERNAME'@'localhost' IDENTIFIED BY '$PASSWORD';
+CREATE DATABASE IF NOT EXISTS nextcloud CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci;
+GRANT ALL PRIVILEGES ON nextcloud.* TO '$USERNAME'@'localhost';
 FLUSH PRIVILEGES;"
 
-#Downloading nextcloud zip file
-apt install -y wget tar
-wget https://download.nextcloud.com/server/releases/latest.tar.bz2
-echo "Unpacking nextcloud..."
-tar -xf latest.tar.bz2
-echo "Done"
-rm latest.tar.bz2
-#Move nextcloud to apache folder
-rm /var/www/html/index.html
-echo "Moving nextcloud to apache folder..."
-mv nextcloud/ /var/www/html/
-echo "Done"
+#Downloading nextcloud tar.bz2 file
+apt install -y wget tar curl
 
+nextcloud_link=$(curl https://nextcloud.com/install/\#instructions-server \
+	| grep -Eo 'https://.+\/releases\/.+\.tar\.bz2"' | sed 's/"//g')
+nextcloud_tmp="/tmp/nextcloud.tar.bz2"
+
+wget "$nextcloud_link" -O "$nextcloud_tmp"
+tar -xf "$nextcloud_tmp"
+
+#Copy nextcloud to apache folder
+rm /var/www/html/index.html
+echo "Copying nextcloud to apache folder..."
+cp -r nextcloud/ /var/www/html/
+rm -v "$nextcloud_tmp"
+echo "Done",
 #Apache config
 cat > /etc/apache2/sites-available/nextcloud.conf <<EOL
 Alias /nextcloud "/var/www/html/nextcloud/"
